Add files via upload
Initial formatting idea. This doesn't implement the buttons yet, but it shows the general idea thus far. We'll have to have it set so that activating buttons will cause the additional menus to open up, but it's getting late and I have work in the morning. I'll see what I can do tomorrow.